chore: refactor ViewCartPage splitting components and fixing hook dependencies
