+++ packages/grant-explorer/src/features/round/DonateToGitcoin/index.ts
@@ -1,0 +1,1 @@
+export { DonateToGitcoin } from "./DonateToGitcoin";
